zng_task::channel

Struct IpcBytesWriter

pub struct IpcBytesWriter { /* private fields */ }

Represents an async IpcBytes writer.

Use IpcBytes::new_writer to start writing.

Implementations

impl IpcBytesWriter

pub async fn finish(self) -> Result<IpcBytes>

Finish writing and move data to a shareable IpcBytes.

pub async fn finish_mut(self) -> Result<IpcBytesMut>

Mode data to an exclusive mutable IpcBytes that can be further modified, but not resized.