Struct zng_task::channel::Receiver

pub struct Receiver<T>(/* private fields */);

The receiving end of a channel.

Use bounded,unbounded or rendezvous to create a channel.

Work Stealing

Cloning the receiver does not turn this channel into a broadcast channel. Each message will only be received by a single receiver. You can use this to to implement work stealing.

Implementations

impl<T> Receiver<T>

pub async fn recv(&self) -> Result<T, RecvError>

Wait for an incoming value from the channel associated with this receiver.

Returns an error if all senders have been dropped.

pub async fn recv_deadline( &self, deadline: impl Into<Deadline> ) -> Result<T, RecvTimeoutError>

Wait for an incoming value from the channel associated with this receiver.

Returns an error if all senders have been dropped or the deadline is reached.

pub fn is_disconnected(&self) -> bool

Returns true if all senders for this channel have been dropped.

pub fn is_empty(&self) -> bool

Returns true if the channel is empty.

Note: rendezvous channels are always empty.

pub fn is_full(&self) -> bool

Returns true if the channel is full.

Note: rendezvous channels are always full and unbounded channels are never full.

pub fn len(&self) -> usize

Returns the number of messages in the channel.

pub fn capacity(&self) -> Option<usize>

If the channel is bounded, returns its capacity.

pub fn drain(&self) -> Drain<'_, T>

Takes all sitting in the channel.