zng_task::channel

Trait IpcValue

pub trait IpcValue:
    Serialize
    + for<'d> Deserialize<'d>
    + Send
    + 'static { }

Represents a type that can be an input and output of IPC channels.

Trait Alias

This trait is used like a type alias for traits and is already implemented for all types it applies to.

Implementing

Types need to be serde::Serialize + serde::de::Deserialize + Send + 'static to auto-implement this trait, if you want to send an external type in that does not implement all the traits you may need to declare a newtype wrapper.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.