zng_task

Function all_some

pub async fn all_some<Some, F: IntoFuture<Output = Option<Some>>>(
    futures: impl IntoIterator<Item = F>,
) -> Option<Vec<Some>>

A future that awaits on all futures at the same time and returns when all futures are Some(_) or any future is None.

This is the dynamic version of all_some!.