zng_task

Function any_some

pub async fn any_some<Some, F: IntoFuture<Output = Option<Some>>>(
    futures: impl IntoIterator<Item = F>,
) -> Option<Some>

A future that awaits on all futures at the same time and returns when any future is Some(_) or all are None.

This is the dynamic version of all_some!.