Function zng_task::doc_test

pub fn doc_test<F>(
    spin: bool,
    task: impl IntoFuture<IntoFuture = F>,
) -> F::Output
where
    F: Future,

Executor used in async doc tests.

If spin is true the spin_on executor is used with a timeout of 500 milliseconds. IF spin is false the block_on executor is used with a timeout of 5 seconds.