zng_task

Function poll_spawn

pub fn poll_spawn<F>(task: impl IntoFuture<IntoFuture = F>)
where
    F: Future<Output = ()> + Send + 'static,

Polls the task once immediately on the calling thread, if the task is pending, continues execution in spawn.