zng_task::http
Function default_client

pub fn default_client() -> &'static Client

The Client used by the functions in this module.

You can replace the default client at the start of the process using set_default_client_init.

Defaults

The default client is created using Client::new.