zng_task::http

Struct ClientBuilder

pub struct ClientBuilder { /* private fields */ }

Builder that can be used to create a Client.

Use Client::builder to start building.

Examples

use zng_task::http::*;

let client = Client::builder().metrics(true).build();

Implementations

impl ClientBuilder

pub fn new() -> Self

New default builder.

pub fn build(self) -> Client

Build the Client using the configured options.

pub fn build_custom<F>(self, custom: F) -> Result<Client, Error>
where F: FnOnce(HttpClientBuilder) -> Result<HttpClient, Error>,

Build the client with more custom build calls in the inner builder.

pub fn default_header( self, key: impl TryHeaderName, value: impl TryHeaderValue, ) -> Result<Self, Error>

Add a default header to be passed with every request.

pub fn cookies(self) -> Self

Enable persistent cookie handling for all requests using this client using a shared cookie jar.

pub fn cookie_jar(self, cookie_jar: CookieJar) -> Self

Set a cookie jar to use to accept, store, and supply cookies for incoming responses and outgoing requests.

Note that the default_client already has a cookie jar.

pub fn timeout(self, timeout: Duration) -> Self

Specify a maximum amount of time that a complete request/response cycle is allowed to take before being aborted. This includes DNS resolution, connecting to the server, writing the request, and reading the response.

Note that this includes the response read operation, so if you get a response but don’t read-it within this timeout you will get a TimedOut IO error.

By default no timeout is used.

pub fn connect_timeout(self, timeout: Duration) -> Self

Set a timeout for establishing connections to a host.

If not set, the default_client default of 90 seconds will be used.

pub fn low_speed_timeout(self, low_speed: u32, timeout: Duration) -> Self

Specify a maximum amount of time where transfer rate can go below a minimum speed limit.

The low_speed limit is in bytes/s. No low-speed limit is configured by default.

pub fn redirect_policy(self, policy: RedirectPolicy) -> Self

Set a policy for automatically following server redirects.

If enabled the “Referer” header will be set automatically too.

pub fn auto_decompress(self, enabled: bool) -> Self

Enable or disable automatic decompression of the response body.

If enabled the “Accept-Encoding” will also be set automatically, if it was not set using default_header.

This is enabled by default.

pub fn max_upload_speed(self, max: u64) -> Self

Set a maximum upload speed for the request body, in bytes per second.

pub fn max_download_speed(self, max: u64) -> Self

Set a maximum download speed for the response body, in bytes per second.

pub fn metrics(self, enable: bool) -> Self

Enable or disable metrics collecting.

When enabled you can get the information using the Response::metrics method.

This is enabled by default.

pub fn cache(self, cache: impl CacheDb) -> Self

Sets the CacheDb to use.

Caching is only enabled if there is a DB, no caching is done by default.

pub fn cache_mode( self, selector: impl Fn(&Request) -> CacheMode + Send + Sync + 'static, ) -> Self

Sets the CacheMode selector.

The selector closure is called for every cacheable request before it is made, it must return a CacheMode value that configures how the cache is used.

Note that the closure is only called if a cache is set.