zng_task::http
Struct Request

pub struct Request { /* private fields */ }

HTTP request.

Use send to send a request.

Implementations

impl Request

pub fn builder() -> RequestBuilder

Starts an empty builder.

Examples

use zng_task::http;

let request = http::Request::builder().method(http::Method::PUT)?.uri("https://httpbin.org/put")?.build();

Call build or body to finish building the request, note that there are is also an associated function to start a builder for each HTTP method and uri.

pub fn get(uri: impl TryUri) -> Result<RequestBuilder, Error>

Starts building a GET request.

Examples

use zng_task::http;

let get = http::Request::get("https://httpbin.org/get")?.build();

pub fn put(uri: impl TryUri) -> Result<RequestBuilder, Error>

Starts building a PUT request.

Examples

use zng_task::http;

let put = http::Request::put("https://httpbin.org/put")?.header("accept", "application/json")?.build();

pub fn post(uri: impl TryUri) -> Result<RequestBuilder, Error>

Starts building a POST request.

Examples

use zng_task::http;

let post = http::Request::post("https://httpbin.org/post")?.header("accept", "application/json")?.build();

pub fn delete(uri: impl TryUri) -> Result<RequestBuilder, Error>

Starts building a DELETE request.

Examples

use zng_task::http;

let delete = http::Request::delete("https://httpbin.org/delete")?.header("accept", "application/json")?.build();

pub fn patch(uri: impl TryUri) -> Result<RequestBuilder, Error>

Starts building a PATCH request.

Examples

use zng_task::http;

let patch = http::Request::patch("https://httpbin.org/patch")?.header("accept", "application/json")?.build();

pub fn head(uri: impl TryUri) -> Result<RequestBuilder, Error>

Starts building a HEAD request.

Examples

use zng_task::http;

let head = http::Request::head("https://httpbin.org")?.build();

pub fn uri(&self) -> &Uri

Returns a reference to the associated URI.

pub fn method(&self) -> &Method

Returns a reference to the associated HTTP method.

pub fn headers(&self) -> &HeaderMap

Returns a reference to the associated header field map.

pub fn clone_with(&self, body: impl TryBody) -> Result<Self, Error>

Create a clone of the request method, URI, version and headers, with a new body.