zng_task

Module io

IO tasks.

Most of the types in this module are re-exported from futures_lite::io.

Re-exports

pub use futures_lite::io::AsyncBufRead;

pub use futures_lite::io::AsyncBufReadExt;

pub use futures_lite::io::AsyncRead;

pub use futures_lite::io::AsyncReadExt;

pub use futures_lite::io::AsyncSeek;

pub use futures_lite::io::AsyncSeekExt;

pub use futures_lite::io::AsyncWrite;

pub use futures_lite::io::AsyncWriteExt;

pub use futures_lite::io::BoxedReader;

pub use futures_lite::io::BoxedWriter;

pub use futures_lite::io::BufReader;

pub use futures_lite::io::BufWriter;

pub use futures_lite::io::Cursor;

pub use futures_lite::io::ReadHalf;

pub use futures_lite::io::WriteHalf;

pub use futures_lite::io::copy;

pub use futures_lite::io::empty;

pub use futures_lite::io::repeat;

pub use futures_lite::io::sink;

pub use futures_lite::io::split;

Structs

CloneableError

Represents the cloneable parts of an Error.

McBufReader

Multiple consumer buffered read.

Measure

Measure read/write of an async task.

Metrics

Information about the state of an async IO task.

ReadLimited

Represents a stream reader that generates an error if the source stream exceeds a limit.

Statics

METRICS_ID

Metrics in a Progress::with_meta metadata.

Traits

McBufErrorExt

Extension methods for std::io::Error to be used with errors returned by McBufReader.