Module zng_task::ipc

IPC tasks.

This module uses ipc_channel and duct crates to define a worker process that can run tasks in a separate process instance.

Each worker process can run multiple tasks in parallel, the worker type is Worker. Note that this module does not offer a fork implementation, the worker processes begin from the start state. The primary use of process tasks is to make otherwise fatal tasks recoverable, if the task calls unsafe code or code that can potentially terminate the entire process it should run using a Worker. If you only want to recover from panics in safe code consider using task::run_catch or task::wait_catch instead.

This module also re-exports some ipc_channel types and functions. You can send IPC channels in the task request messages, this can be useful for implementing progress reporting or to transfer large byte blobs.

Examples

The example below demonstrates a worker-process setup that uses the same executable as the app-process.

fn main() {
    // this must be called before the app start, when the process
    // is a worker this function never returns.
    task::ipc::run_worker(worker);

    // normal app init..
}

// All tasks for the same worker-process instance must be defined on the same type.
#[derive(Debug, serde::Serialize, serde::Deserialize)]
enum IpcRequest {
    Task1,
    Task2,
}
#[derive(Debug, serde::Serialize, serde::Deserialize)]
enum IpcResponse {
    Result1,
    Result2,
}

// This handler is called for every worker task, in the worker-process.
async fn worker(args: task::ipc::RequestArgs<IpcRequest>) -> IpcResponse {
   println!("received request `{:?}` in worker-process #{}", &args.request, std::process::id());
    match args.request {
        IpcRequest::Task1 => IpcResponse::Result1,
        IpcRequest::Task2 => IpcResponse::Result2,
    }
}

// This runs in the app-process, it starts a worker process and requests a task run.
async fn on_click() {
    println!("app-process #{} starting a worker", std::process::id());
    let mut worker = match task::ipc::Worker::start().await {
        Ok(w) => w,
        Err(e) => {
            eprintln!("error: {e}");
            return;
        },
    };
    // request a task run and await it.
    match worker.run(IpcRequest::Task1).await {
        Ok(IpcResponse::Result1) => println!("ok."),
        Err(e) => eprintln!("error: {e}"),
        _ => unreachable!(),
    }
    // multiple tasks can be requested in parallel, use `task::all!` to await ..

    // the worker process can be gracefully shutdown, awaits all pending tasks.
    let _ = worker.shutdown().await;
}

Note that you can setup different worker types on the same executable using Worker::start_with with a custom environment variable that switches the run_worker call.

fn run_workers() {
    match std::env::var("MY_APP_WORKER") {
        Ok(name) => match name.as_str() {
            "worker_a" => task::ipc::run_worker(worker_a),
            "worker_b" => task::ipc::run_worker(worker_b),
            unknown => panic!("unknown worker, {unknown:?}"),
        },
        Err(e) => match e {
            std::env::VarError::NotPresent => {} // not a worker run
            e => panic!("invalid worker name, {e}"),
        },
    }
}

async fn worker_a(args: task::ipc::RequestArgs<bool>) -> char {
    if args.request {
        'A'
    } else {
        'a'
    }
}

async fn worker_b(args: task::ipc::RequestArgs<char>) -> bool {
    args.request == 'B' || args.request == 'b'
}

fn main() {
    self::run_workers();

    // normal app init..
}

// And in the app side:
async fn on_click() {
    let mut worker_a = task::ipc::Worker::start_with(&[("MY_APP_WORKER", "worker_a")], &[]).await.unwrap();
    let r = worker_a.run(true).await.ok();
    assert_eq!(r, Some('A'));
}

Re-exports

pub use ipc_channel::ipc::bytes_channel;

pub use ipc_channel::ipc::IpcBytesReceiver;

pub use ipc_channel::ipc::IpcBytesSender;

pub use ipc_channel::ipc::IpcReceiver;

pub use ipc_channel::ipc::IpcSender;

Structs

RequestArgs

Arguments for run_worker.

Worker

Represents a running worker process.

WorkerCrashError

Info about a worker process crash.

Enums

RunError

Worker run error.

Constants

VERSION

The App Process and Worker Process must be build using the same exact version and this is validated during run-time, causing a panic if the versions don’t match.

Traits

IpcValue

Represents a type that can be an input and output of IPC workers.

Functions

run_worker

If the process was started by a Worker runs the worker loop and never returns. If not started as worker does nothing.