zng_task

Module parking_lot

Recommended blocking locks.

The Mutex and RwLock types reexported here are recommended, they are more optimized than std alternatives because they don’t implement lock poisoning, have const initialization and integrate with the zng feature "deadlock_detection".

Full API

See the parking_lot crate for the full API.

Re-exports

pub use ::parking_lot::ArcMutexGuard;

pub use ::parking_lot::ArcRwLockReadGuard;

pub use ::parking_lot::ArcRwLockWriteGuard;

pub use ::parking_lot::Condvar;

pub use ::parking_lot::MappedMutexGuard;

pub use ::parking_lot::MappedRwLockReadGuard;

pub use ::parking_lot::MappedRwLockWriteGuard;

pub use ::parking_lot::Mutex;

pub use ::parking_lot::MutexGuard;

pub use ::parking_lot::RwLock;

pub use ::parking_lot::RwLockReadGuard;

pub use ::parking_lot::RwLockUpgradableReadGuard;

pub use ::parking_lot::RwLockWriteGuard;