zng_task::process::worker

Struct Worker

pub struct Worker<I: IpcValue, O: IpcValue> { /* private fields */ }

Represents a running worker process.

Implementations

impl<I: IpcValue, O: IpcValue> Worker<I, O>

pub async fn start(worker_name: impl Into<Txt>) -> Result<Self>

Start a worker process implemented in the current executable.

Note that the current process must call run_worker at startup to actually work. You can use zng_env::on_process_start! to inject startup code.

pub async fn start_with( worker_name: impl Into<Txt>, env_vars: &[(&str, &str)], args: &[&str], ) -> Result<Self>

Start a worker process implemented in the current executable with custom env vars and args.

pub async fn start_other( worker_name: impl Into<Txt>, worker_exe: impl Into<PathBuf>, env_vars: &[(&str, &str)], args: &[&str], ) -> Result<Self>

Start a worker process implemented in another executable with custom env vars and args.

pub async fn shutdown(self) -> Result<()>

Awaits current tasks and kills the worker process.

pub fn run( &mut self, input: I, ) -> impl Future<Output = Result<O, RunError>> + Send + 'static

Run a task in a free worker process thread.

pub fn crash_error(&mut self) -> Option<&WorkerCrashError>

Reference the crash error.

The worker cannot be used if this is set, run requests will immediately disconnect.