zng_task

Module rayon

Parallel iterators.

This module mostly reexports the primary traits from the rayon crate.

This module also includes ParallelIteratorWithCtx that propagates the zng app context to rayon tasks.

Full API

See the rayon crate for the full API.

Re-exports

pub use ::rayon::iter;

pub use ::rayon::slice;

pub use ::rayon::str;

Modules

prelude

Rayon traits imported as _.

Structs

ParallelIteratorWithCtx

Parallel iterator adapter the propagates the thread context.

Traits

ParallelIteratorExt

Extends rayon’s ParallelIterator with thread context.