Function zng_task::wait_respond

pub fn wait_respond<R, F>(task: F) -> ResponseVar<R>
where
    R: VarValue,
    F: FnOnce() -> R + Send + 'static,

Like spawn_wait, but the task will send its result to a ResponseVar<R>.

Cancellation

Dropping the ResponseVar<R> does not cancel the task, it will still run to completion.

Panic Handling

If the task panics the panic is logged as an error and resumed in the response var modify closure.